- Changed the paragraph detection implementation. That was kidda funny, because the algorithm is pretty much the same. It seems PHP has some idiosycrancies with references and scope resolution. Well, at least the problem has been settle down, gone for good.
- Changed the private record char detection. Actually, I just exchanged the {}, used
to choose a char within a string, fro a [] (as recommended for PHP 6 compatibility).

- Improved the diary's entries selection algorithm. It will try to get the last seven
days or at least seven entries, whatever came first. You could imagine this would be
an infinite loop for a small database. Yeah, indeed. So I established a hard limit,
of 60 days. PHP will still loops a lot, but...